Extract OpenAIBatchCompletionService to a separate file
Moved OpenAIBatchCompletionService from llm_service.py to a new file llm_batch_service.py. This refactor improves code modularity and separation of concerns, making the codebase easier to maintain.
