+++ lib/llm/llm_service.py
@@ -1,6 +1,4 @@
-import json
 import os
-import secrets
 from abc import ABC, abstractmethod
 
 import tiktoken
@@ -13,12 +11,11 @@ from langchain_chroma import Chroma
 from langchain_openai import ChatOpenAI
 from langchain_openai import OpenAIEmbeddings
 from openai import OpenAI
-from openai.types import ImagesResponse, Batch
+from openai.types import ImagesResponse
 from openai.types.chat import ChatCompletion
 
 from config.settings import MEDIA_ROOT
-from lib.llm.valueobject.chat import Message, RoleType
-from lib.llm.valueobject.chat_batch import MessageChunk
+from lib.llm.valueobject.chat import Message
 from lib.llm.valueobject.config import OpenAIGptConfig, GeminiConfig
 from lib.llm.valueobject.rag import PdfDataloader
 from llm_chat.domain.valueobject.chat import MessageDTO
@@ -206,131 +203,3 @@ class OpenAILlmRagService(LlmService):
         )
 
         return chain.invoke({"question": message.content})
-
-
-class OpenAIBatchCompletionService(LlmService):
-    def __init__(self, config: OpenAIGptConfig):
-        super().__init__()
-        self.config = config
-
-    def create_message_chunk(self, messages: list[Message]) -> MessageChunk:
-        return MessageChunk(
-            messages=messages,
-            model=self.config.model,
-            max_tokens=self.config.max_tokens,
-        )
-
-    @staticmethod
-    def export_jsonl_file(chunks: list[MessageChunk]) -> str:
-        """
-        指定された MessageChunk リストを JSONL 形式に変換して、ローカルファイルとして保存します。
-
-        JSONL 形式の各行には、1つのチャンク（MessageChunk）がシリアライズされた
-        JSON エントリが含まれています。
-
-        Args:
-            chunks (list[MessageChunk]): JSONL形式に変換する対象のデータチャンクリスト。
-
-        Returns:
-            str: 作成された JSONL ファイルの絶対パス。
-
-        Raises:
-            RuntimeError: ファイルの作成やデータの書き込み中にエラーが発生した場合。
-
-        Example:
-            >>> chunks1 = [
-            >>>     MessageChunk(...),
-            >>>     MessageChunk(...),
-            >>> ]
-            >>> file_path1 = OpenAIBatchCompletionService.export_jsonl_file(chunks1)
-            >>> print(f"File saved at: {file_path}")
-
-        Note:
-            作成されたファイルは一時的なもので、後続の処理が完了後に削除されることが想定されています。
-        """
-        file_name = f"export_{secrets.token_hex(5)}.jsonl"
-        file_path = os.path.abspath(file_name)
-        try:
-            with open(file_path, "w", encoding="utf-8") as jsonl_file:
-                for chunk in chunks:
-                    json_entry = chunk.to_jsonl_entry()
-                    jsonl_file.write(json.dumps(json_entry) + "\n")
-        except Exception as e:
-            raise RuntimeError(f"Failed to export JSONL file: {str(e)}")
-
-        return file_path
-
-    def upload_jsonl_file(self, chunks: list[MessageChunk]) -> str:
-        """
-        JSONLファイルをOpenAIにアップロードする。
-
-        Args:
-            chunks (list[MessageChunk]): アップロードするデータのチャンク。
-
-        Returns:
-            str: アップロードされたファイルのID。
-        """
-        jsonl_file_path = self.export_jsonl_file(chunks)
-
-        try:
-            with open(jsonl_file_path, "rb") as jsonl_file:
-                uploaded_file = OpenAI(api_key=self.config.api_key).files.create(
-                    file=jsonl_file, purpose="batch"
-                )
-            return uploaded_file.id
-        finally:
-            if os.path.exists(jsonl_file_path):
-                os.remove(jsonl_file_path)
-
-    def create_batch(self, uploaded_file_id: str) -> Batch:
-        return OpenAI(api_key=self.config.api_key).batches.create(
-            input_file_id=uploaded_file_id,
-            endpoint="/v1/chat/completions",
-            completion_window="24h",
-        )
-
-    def retrieve_answer(self, batch_id: str) -> Batch:
-        return OpenAI(api_key=self.config.api_key).batches.retrieve(batch_id)
-
-    @staticmethod
-    def parse_to_messages(results: list[dict]) -> list[Message]:
-        """
-        OpenAIBatchCompletionService.retrieve_content で取得された結果を
-        Message クラスのインスタンスリストに変換する。
-
-        Args:
-            results (list[dict]): OpenAIから取得した結果のリスト。
-
-        Returns:
-            list[Message]: Messageインスタンスのリスト。
-        """
-        messages = []
-        for result in results:
-            try:
-                choice = result["response"]["body"]["choices"][0]
-                messages.append(
-                    Message(
-                        role=RoleType(choice["message"]["role"].upper()),
-                        content=choice["message"]["content"],
-                    )
-                )
-            except (KeyError, ValueError) as e:
-                print(f"Error parsing result: {result}, error: {e}")
-        return messages
-
-    def retrieve_content(self, file_id: str) -> list[Message]:
-        raw_data = OpenAI(api_key=self.config.api_key).files.content(file_id).content
-        file_name = f"retrieve_{secrets.token_hex(5)}.jsonl"
-        file_path = os.path.abspath(file_name)
-        with open(file_path, "wb") as file:
-            file.write(raw_data)
-
-        results = []
-        with open(file_path, "r") as file:
-            for line in file:
-                # Parsing the JSON string into a dict and appending to the list of results
-                json_object = json.loads(line.strip())
-                results.append(json_object)
-        print(f"{results=}")
-        results = self.parse_to_messages(results)
-        return results
